@@ -39,23 +39,23 @@ dependencies:
       url: https://github.com/effektio/flyer-flutter_chat_ui.git
   flutter_chat_types: ^3.3.2
   google_fonts: ^2.3.1
-  flutter_svg: ^0.22.0
+  flutter_svg: ^1.1.6
   duration: ^3.0.11
-  file_picker: ^4.5.1
-  open_file: ^3.2.1
+  file_picker: ^5.2.5
+  open_app_file: ^4.0.1
   cached_network_image: ^3.2.1
   flutter_link_previewer: ^3.0.1
   bubble: ^1.2.1
   colorize_text_avatar: ^1.0.2
-  html_editor_enhanced: ^2.5.0
+  html_editor_enhanced: ^2.5.1
   themed: ^4.0.0
   effektio_flutter_sdk:
     path: packages/rust_sdk
   mime: ^1.0.2
   flutter_colorpicker: ^1.0.3
-  emoji_picker_flutter: ^1.3.0
+  emoji_picker_flutter: ^1.5.2
   filesystem_picker: ^2.0.1
-  permission_handler: ^9.2.0
+  permission_handler: ^10.2.0
   transparent_image: ^2.0.0
   flutter_slidable: ^1.3.0
   expandable_text: ^2.3.0
@@ -63,13 +63,14 @@ dependencies:
   characters: ^1.2.0
   sprintf: ^6.0.0
   photo_view: ^0.14.0
-  flutter_parsed_text: ^2.2.1
   avatar_stack: ^1.1.2
   flutter_lorem: ^2.0.0
   get_time_ago: ^1.1.5
-  implicitly_animated_reorderable_list: ^0.4.2
+  implicitly_animated_reorderable_list_2: ^0.5.0
   flutter_mentions: ^2.0.1
-  flutter_matrix_html: ^1.0.0
+  flutter_matrix_html:
+    git:
+      url: https://github.com/effektio/flutter_matrix_html.git
   flutter_icons_null_safety: ^1.1.0
   connectivity_plus: ^3.0.2
   overlay_support: ^2.1.0
