@@ -11,8 +11,7 @@ import 'package:effektio/screens/HomeScreens/chat/Overview.dart';
 // import 'package:effektio/screens/HomeScreens/Notification.dart';
 import 'package:effektio/screens/HomeScreens/faq/Overview.dart';
 import 'package:effektio/screens/HomeScreens/news/News.dart';
-import 'package:effektio/screens/HomeScreens/todo/AddToDo.dart';
-import 'package:effektio/screens/HomeScreens/todo/ToDoScreen.dart';
+import 'package:effektio/screens/HomeScreens/todo/ToDo.dart';
 import 'package:effektio/screens/OnboardingScreens/LogIn.dart';
 import 'package:effektio/screens/OnboardingScreens/Signup.dart';
 import 'package:effektio/screens/SideMenuScreens/Gallery.dart';
@@ -86,7 +85,6 @@ class Effektio extends StatelessWidget {
               '/profile': (BuildContext context) => const SocialProfileScreen(),
               '/signup': (BuildContext context) => const SignupScreen(),
               '/gallery': (BuildContext context) => const GalleryScreen(),
-              '/addTodo': (BuildContext context) => const AddToDoScreen(),
             },
           ),
         ),
@@ -309,19 +307,19 @@ class _EffektioHomeState extends State<EffektioHome>
             userId: client.userId().toString(),
             displayName: displayName,
             displayAvatar: displayAvatar,
-            client: client,
           ),
           bottomNavigationBar: TabBar(
             labelColor: AppCommonTheme.primaryColor,
             unselectedLabelColor: AppCommonTheme.svgIconColor,
             controller: tabController,
+            indicatorSize: TabBarIndicatorSize.tab,
+            indicatorPadding: const EdgeInsets.symmetric(horizontal: 12),
             indicator: const MaterialIndicator(
               height: 5,
               bottomLeftRadius: 8,
               bottomRightRadius: 8,
               topLeftRadius: 0,
               topRightRadius: 0,
-              horizontalPadding: 12,
               tabPosition: TabPosition.top,
               color: AppCommonTheme.primaryColor,
             ),
