@@ -1,5 +1,5 @@
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart'
-    show TaskUpdateBuilder;
+    show CommentsManager, TaskUpdateBuilder;
 import 'package:flutter/widgets.dart';
 
 //Todo task model.
@@ -8,6 +8,7 @@ class ToDoTask {
   final String name;
   final String? description;
   final TaskUpdateBuilder taskUpdateDraft;
+  final CommentsManager commentsManager;
   final Color? color;
   final List<String>? assignees;
   final List<String>? categories;
@@ -18,11 +19,12 @@ class ToDoTask {
   final DateTime? start;
   final DateTime? due;
 
-  ToDoTask({
+  const ToDoTask({
     this.index,
     required this.name,
     required this.progressPercent,
     required this.taskUpdateDraft,
+    required this.commentsManager,
     this.assignees,
     this.categories,
     this.tags,
@@ -40,6 +42,7 @@ class ToDoTask {
     required String name,
     required TaskUpdateBuilder taskUpdateDraft,
     required int progressPercent,
+    required CommentsManager commentsManager,
     String? description,
     Color? color,
     List<String>? subscribers,
@@ -54,6 +57,7 @@ class ToDoTask {
       index: index ?? this.index,
       name: name,
       taskUpdateDraft: taskUpdateDraft,
+      commentsManager: commentsManager,
       progressPercent: progressPercent,
       description: description ?? this.description,
       color: color ?? this.color,
