@@ -51,14 +51,12 @@ class _ChatScreenState extends State<ChatScreen> {
   @override
   void initState() {
     super.initState();
-
     roomController.setCurrentRoom(widget.room);
   }
 
   @override
   void dispose() {
     roomController.setCurrentRoom(null);
-
     super.dispose();
   }
 
@@ -146,6 +144,7 @@ class _ChatScreenState extends State<ChatScreen> {
       builder: (ChatRoomController controller) {
         if (!controller.isEmojiContainerVisible) {
           return CustomChatInput(
+            roomController: controller,
             isChatScreen: true,
             roomName: widget.roomName ?? AppLocalizations.of(context)!.noName,
             onButtonPressed: () => onSendButtonPressed(controller),
@@ -293,7 +292,6 @@ class _ChatScreenState extends State<ChatScreen> {
       message: p1,
       onPreviewDataFetched: roomController.handlePreviewDataFetched,
       messageWidth: messageWidth,
-      controller: roomController,
     );
   }
 
