@@ -1,7 +1,7 @@
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
-import 'package:effektio/screens/HomeScreens/todo/screens/MyAssignments.dart';
-import 'package:effektio/screens/HomeScreens/todo/screens/MyRecentActivity.dart';
-import 'package:effektio/screens/HomeScreens/todo/screens/ToDoBookmarks.dart';
+import 'package:effektio/screens/HomeScreens/todo/MyAssignments.dart';
+import 'package:effektio/screens/HomeScreens/todo/MyRecentActivity.dart';
+import 'package:effektio/screens/HomeScreens/todo/ToDoBookmarks.dart';
 import 'package:effektio/widgets/AppCommon.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_icons_null_safety/flutter_icons_null_safety.dart';
@@ -70,8 +70,11 @@ class _ToDoMineScreenState extends State<ToDoMineScreen> {
           ),
         ),
         ListTile(
-          onTap: (){
-            showNotYetImplementedMsg(context, 'Upcoming events is not yet Implemented');
+          onTap: () {
+            showNotYetImplementedMsg(
+              context,
+              'Upcoming events is not yet Implemented',
+            );
           },
           leading: const Icon(
             FlutterIcons.event_mdi,
