@@ -21,8 +21,8 @@
  */
 
 import 'package:effektio/common/constants.dart';
-import 'package:effektio/widgets/emoji_picker_widget.dart';
-import 'package:effektio/widgets/reaction_popup_configuration.dart';
+import 'package:effektio/widgets/EmojiPicker.dart';
+import 'package:effektio/widgets/ReactionPopUpConfig.dart';
 import 'package:expandable_text/expandable_text.dart';
 import 'package:flutter/material.dart';
 
