@@ -9,122 +9,113 @@ import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 import 'package:flutter_link_previewer/flutter_link_previewer.dart';
 import 'package:flutter_matrix_html/flutter_html.dart';
+import 'package:get/get.dart';
 
-class TextMessageBuilder extends StatefulWidget {
-  const TextMessageBuilder({
+class TextMessageBuilder extends StatelessWidget {
+  TextMessageBuilder({
     Key? key,
     required this.message,
     this.onPreviewDataFetched,
-    required this.controller,
     required this.messageWidth,
   }) : super(key: key);
 
   final types.TextMessage message;
   final void Function(types.TextMessage, types.PreviewData)?
       onPreviewDataFetched;
-  final ChatRoomController controller;
   final int messageWidth;
-
-  @override
-  State<TextMessageBuilder> createState() => _TextMessageBuilderState();
-}
-
-class _TextMessageBuilderState extends State<TextMessageBuilder> {
-  bool isLoading = true;
-
-  void toggleLoading() {
-    setState(() {
-      isLoading = !isLoading;
-    });
-  }
-
+  final controller = Get.find<ChatRoomController>();
   @override
   Widget build(BuildContext context) {
     //remove mx-reply tags.
-    String parsedString = simplifyBody(widget.message.text);
+    String parsedString = simplifyBody(message.text);
     final urlRegexp = RegExp(
       r'https://matrix\.to/#/@[A-Za-z0-9]+:[A-Za-z0-9]+\.[A-Za-z0-9]+',
       caseSensitive: false,
     );
-    final bool enlargeEmoji = widget.message.metadata!['enlargeEmoji'];
+    final bool enlargeEmoji = message.metadata!['enlargeEmoji'];
     //will return empty if link is other than mention
     final matches = urlRegexp.allMatches(parsedString);
     if (matches.isEmpty) {
-      return _linkPreview(
-        widget.message.author,
-        widget.messageWidth.toDouble(),
-        context,
-        parsedString,
+      return LinkPreview(
+        metadataTitleStyle: controller.userId == message.author.id
+            ? const EffektioChatTheme().sentMessageLinkTitleTextStyle
+            : const EffektioChatTheme().receivedMessageLinkTitleTextStyle,
+        metadataTextStyle: controller.userId == message.author.id
+            ? const EffektioChatTheme().sentMessageLinkDescriptionTextStyle
+            : const EffektioChatTheme().receivedMessageLinkDescriptionTextStyle,
+        enableAnimation: true,
+        imageBuilder: (image) {
+          return Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: ClipRRect(
+              borderRadius: BorderRadius.circular(16),
+              child: CachedNetworkImage(
+                imageUrl: image,
+                maxHeightDiskCache: 256,
+              ),
+            ),
+          );
+        },
+        previewData: message.previewData,
+        text: parsedString,
+        onPreviewDataFetched: _onPreviewDataFetched,
+        textWidget: _TextWidget(
+          controller: controller,
+          message: message,
+          enlargeEmoji: message.metadata!['enlargeEmoji'],
+        ),
+        width: messageWidth.toDouble(),
+        padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 16),
       );
     }
-    return textWidget(enlargeEmoji);
+    return _TextWidget(
+      controller: controller,
+      message: message,
+      enlargeEmoji: enlargeEmoji,
+    );
   }
 
-  Widget textWidget(bool enlargeEmoji) {
-    final emojiTextStyle = widget.controller.userId == widget.message.author.id
+  void _onPreviewDataFetched(types.PreviewData previewData) {
+    final controller = Get.find<ChatRoomController>();
+    if (message.previewData == null) {
+      controller.handlePreviewDataFetched.call(message, previewData);
+    }
+  }
+}
+
+class _TextWidget extends StatelessWidget {
+  const _TextWidget({
+    required this.controller,
+    required this.message,
+    required this.enlargeEmoji,
+  });
+  final ChatRoomController controller;
+  final types.TextMessage message;
+  final bool enlargeEmoji;
+  @override
+  Widget build(BuildContext context) {
+    final emojiTextStyle = controller.userId == message.author.id
         ? const EffektioChatTheme().sentEmojiMessageTextStyle
         : const EffektioChatTheme().receivedEmojiMessageTextStyle;
     return ConstrainedBox(
       constraints: BoxConstraints(
         maxWidth: enlargeEmoji
             ? double.infinity
-            : sqrt(widget.message.metadata!['messageLength']) * 38.5,
+            : sqrt(message.metadata!['messageLength']) * 38.5,
         maxHeight: double.infinity,
       ),
       child: enlargeEmoji
           ? Text(
-              widget.message.text,
+              message.text,
               style: emojiTextStyle,
             )
           : Html(
               // ignore: prefer_single_quotes, unnecessary_string_interpolations
-              data: """${widget.message.text}""",
+              data: """${message.text}""",
               padding: const EdgeInsets.all(8),
               defaultTextStyle:
                   const TextStyle(color: ChatTheme01.chatBodyTextColor),
             ),
     );
-  }
-
-  Widget _linkPreview(
-    types.User user,
-    double width,
-    BuildContext context,
-    String parsedString,
-  ) {
-    return LinkPreview(
-      metadataTitleStyle: user.id == widget.message.author.id
-          ? const EffektioChatTheme().sentMessageLinkTitleTextStyle
-          : const EffektioChatTheme().receivedMessageLinkTitleTextStyle,
-      metadataTextStyle: user.id == widget.message.author.id
-          ? const EffektioChatTheme().sentMessageLinkDescriptionTextStyle
-          : const EffektioChatTheme().receivedMessageLinkDescriptionTextStyle,
-      enableAnimation: true,
-      imageBuilder: (image) {
-        return Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: ClipRRect(
-            borderRadius: BorderRadius.circular(16),
-            child: CachedNetworkImage(
-              imageUrl: image,
-              maxHeightDiskCache: 256,
-            ),
-          ),
-        );
-      },
-      previewData: widget.message.previewData,
-      text: parsedString,
-      onPreviewDataFetched: _onPreviewDataFetched,
-      textWidget: textWidget(widget.message.metadata!['enlargeEmoji']),
-      width: width,
-      padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 16),
-    );
-  }
-
-  void _onPreviewDataFetched(types.PreviewData previewData) {
-    if (widget.message.previewData == null) {
-      widget.controller.handlePreviewDataFetched
-          .call(widget.message, previewData);
-    }
   }
 }
